fix(dream): 抽取图片下载与 get_bot 容错
